Fix incrementAndInvoice function
Calling incrementAndInvoice without an argument would result in incrementQuantity being called with a null value, which is unexpected behaviour in my mind. Updating incrementAndInvoice to accept a count argument makes it function like the other methods and is thus much more intuitive.

Secondly, this function wasn't actually invoicing the user as one is led to believe by the function name.