Change order of boolean test for faster tests
The logic of the Boolean expression in the previous order makes it so that the event existence is always checked against the Stripe servers first before checking whether this is a test. This results in slower tests as each test will make this round trip even though the condition will make it so the result is irrelevant.

The proposed ordering omits the check against the Stripe servers entirely during tests. This change has no real consequence in terms of effective logic since even if one explicitly wanted it to check that the event exists in Stripe, the `&& !` was making the result irrelevant when testing.