Fix invoicing subscriptions
When invoicing the customer while swapping plans, Cashier would invoice the customer for all its pending items instead of just the ones for the subscription. This PR fixes this by using the subscription field when creating an invoice to allow swapping a plan and the `incrementAndInvoice` method so the invoice at hand only invoices for the subscription and also uses the proper tax which is applied to the subscription.