Handle incomplete subscriptions upon creation
Stripe changed the way failed payments for subscriptions are handled in their 2019-03-14 API update: https://stripe.com/docs/upgrades#2019-03-14

This means subscriptions still get created if the payment fails but they get an incomplete status which gives the customer 23 hours to pay. Since Cashier expects a subscription to be properly paid for we should handle this scenario gracefully.

In the previous API version a card exception was thrown when a payment failed so the only thing that changes with this change is that it's a different exception that's thrown. What also happens is that the subscription gets cancelled immediately if we detect an incomplete subscription so the invoice is voided and the customer isn't billed anymore. No subscription gets added to the billable entity so the developer implementing the feature only needs to handle the exception.