Prevent exception throwing on some methods
The idea behind these changes is that these methods should still be able to be used even if the Billable model isn't a Stripe customer yet. By default we can assume that when a Billable model isn't a Stripe customer yet that no invoices, payment methods or upcoming invoice exists yet.